aider: Changed the domain of the testNet nodes to 'localhost' in src/rchain-networks.js.
# Context:
USER: explain who the node0 appears in the interface.
ASSISTANT: The interface of the application is not included in the provided code. The provided code is a JavaScript file that contains functions for interacting with an RNode server, including sending deploys, fetching data for a deploy, and proposing blocks.

The term "node0" does not appear in the provided code. If "node0" is a term used in the interface of your application, it might refer to a specific instance of an RNode server.

If you want to modify how "node0" is displayed or behaves in the interface, you would need to modify the code that generates the interface. This code might be located in a different JavaScript file, or in an HTML, CSS, or other type of file, depending on how your application is structured.

Please provide the file and the specific changes you want to make, and I will be able to assist you further.
USER: I added these *read-write* files: src/rchain-networks.js
ASSISTANT: I see that you've added the file `src/rchain-networks.js` to the list of *read-write* files. However, I still need more specific information about the changes you want to make.

The file `src/rchain-networks.js` contains definitions for different RChain networks (localNet, testNet, mainNet) and a function `getNodeUrls` to generate URLs for interacting with these networks.

If "node0" is a term used in the interface of your application, it might refer to the first host in the `hosts` array of a network. For example, in the `localNet` network, "node0" could refer to the host with the domain 'localhost' and the ports defined by `defaultPorts`.

Please provide more specific details about the changes you want to make related to "node0".
USER: explain who the node0 appears in the interface.
ASSISTANT: The term "node0" likely refers to the first node in the network configuration for each network (localNet, testNet, mainNet) defined in the `src/rchain-networks.js` file.

In the context of the `localNet` configuration, "node0" would refer to the first object in the `hosts` array:

```javascript
{ domain: 'localhost', shardId: defaultShardId, ...defaultPorts }
```

For the `testNet` and `mainNet` configurations, "node0" would refer to the first node generated by the `getTestNetUrls` and `getMainNetUrls` functions respectively.

In the user interface, "node0" would likely be used to allow the user to select this specific node for interaction. The exact way it appears would depend on the code in the user interface, which is not provided here.

If you want to change how "node0" is displayed or behaves in the interface, you would need to modify the code that generates the interface. This code might be located in a different JavaScript file, or in an HTML, CSS, or other type of file, depending on how your application is structured.

Please provide the file and the specific changes you want to make, and I will be able to assist you further.
USER: how can we modify the code to use localhost for the node0 and other node dropdown definitions
ASSISTANT: To modify the code to use 'localhost' for the node0 and other node dropdown definitions, you would need to change the `domain` property of the objects in the `hosts` array for each network configuration in the `src/rchain-networks.js` file.

Here is an example of how you could modify the `testNet` configuration to use 'localhost' for all nodes:

```javascript
src/rchain-networks.js
<<<<<<< ORIGINAL
const getTestNetUrls = n => {
  const instance = `node${n}`
  return {
    domain: `${instance}.testnet.rchain.coop`,
    instance,
    shardId: testNetShardId,
    ...defaultPortsSSL,
  }
}
=======
const getTestNetUrls = n => {
  const instance = `node${n}`
  return {
    domain: 'localhost',
    instance,
    shardId: testNetShardId,
    ...defaultPortsSSL,
  }
}
>>>>>>> UPDATED
```

This change will make all nodes in the `testNet` configuration use 'localhost' as their domain. You would need to make similar changes for the `mainNet` configuration and any other network configurations you have.

Please note that this change will only affect the domain used for the nodes. The ports and other properties of the nodes will remain the same. If you want to use different ports for each node, you would need to modify the `...defaultPortsSSL` part of the node configuration.

Also, please be aware that using 'localhost' for all nodes might not be suitable for a production environment. This change is more appropriate for a local testing environment.