aider: Added .env file to .gitignore.
# Context:
USER: do not add the .env file to the repo. It should be gitignored
ASSISTANT: You're correct. The `.env` file should not be added to the repository. It should be included in the `.gitignore` file to prevent it from being tracked by Git.

Here's the change you need to make:

```gitignore
.gitignore
<<<<<<< ORIGINAL

=======
.env
>>>>>>> UPDATED
```

After making this change, run `git rm --cached .env` to remove the `.env` file from the repository. This will not delete the file from your local file system, it will just untrack the file in Git.